requirements.txt
pandas>=2.0.3
requests>=2.31.0
apache-airflow>=2.6.3
numpy>=1.25.2
azure-storage-file-datalake>=12.12.0
azure-identity>=1.14.0